aes-gcm-siv: Parallel AES-CTR support
Adds an internal CTR implementation which uses a 32-bit little endian
counter.

As implemented it operates on 8 blocks at a time, allowing for parallel
pipelining of AES-NI.

Criterion benchmarks show this brings the performance down to ~6cpb for
encryption and ~10cpb for decryption, which is at least in the right
ballpark for a performant implementation.