Use `copy_within` for `aes-siv`/`xsalsa20poly1305`; MSRV 1.37+
`aes-siv` and `xsalsa20poly1305` use a prepended authentication tag,
which means during in-place encryption/decryption, they need the message
shifted by the tag size.

It would be nice if this could happen during encryption/decryption:

https://github.com/RustCrypto/block-ciphers/issues/59

...but barring that, using `copy_within` (which performs a `memmove`
behind the scenes) should be the fastest available option.

This commit replaces a previous `for` loop which performed this copy
with `copy_within`.

Note that `copy_within` is MSRV 1.37+, so this bumps the project MSRV
up by a single version (from 1.36).