aes-gcm-siv: add NCC audit notes to docs
A recent audit by NCC group covered the dependencies shared between the
`aes-gcm` and `aes-gcm-siv` crates, namely the `aes` crate (and its
dependencies) as well as the `polyval` crate, which is used by the
`aes-gcm` crate to implement GHASH:

https://research.nccgroup.com/2020/02/26/public-report-rustcrypto-aes-gcm-and-chacha20poly1305-implementation-review/

Note that the `aes-gcm-siv` crate itself was *NOT* audited.

This commit updates the README.md files and crate documentation for
the `aes-gcm-siv` to note that the `aes` crate and `polyval` crate were
audited, but that the `aes-gcm-siv` crate itself has *NOT* been audited.

It also updates the `maintenance` badge from `experimental` to
`actively-developed`.