crypto_box: fix XChaCha20Poly1305 key derivation [SECURITY] (#295)
Previous versions of this crate did not derive a uniformly random key
when used with the XChaCha20Poly1305 AEAD.

The implementation has been updated to use the HChaCha20 function for
this purpose, however this change makes the Curve25519XChaCha20Poly1305
construction implemented by versions earlier than this commit
incompatible.

This was an unfortunate oversight in what is otherwise a non-standard
construction with no other known implementations to test against.
Perhaps libsodium implements this construction and we could try to
ensure compatibility there going forward.

The suggestion for any users of the previous implementation is to
migrate all ciphertexts to the new construction, as the original
construction was broken. Please open an issue if this actually impacts
you and you would like help addressing the problem.

This issue likely deserves a RUSTSEC advisory.