chacha20poly1305: add v0.9.1 release notes
Adds release notes for a v0.9.1 release of `chacha20poly1305` (done on a
separate branch since `master` is curently v0.10.0-pre.1)

This release unpins `zeroize`, allowing any version to be used. It was
pinned to prevent MSRV breakages, however this has proven to be quite
problematic and prevents compatible versions of crates from being used.