Enable `getrandom` feature by defaut; improve docs
The `getrandom` feature transitively activates `aead/getrandom` and
makes `aead::OsRng` available.

RNG access is pretty essential for successful use of these crates. It's
needed to generate random keys as the very least.

The documentation has also been updated to show the generation of random
keys, as opposed to using a hardcoded one.