ocb3: use `Ocb3` as the type name
Renames `AesOcb3` to `Ocb3`.

Unlike AES-GCM, AES-GCM-SIV, and AES-SIV, in which AES is the de facto
cipher for that mode, OCB is defined in a more cipher-agnostic way
similar to CCM or EAX modes (which, as it were, is reflected in our
choice of crate names).

This renames the type to reflect that, and as it were, match the crate
name as well.