CI: fix security audit
New Ubuntu images are being deployed, which is causing the cached
binaries not to work due to a GLIBC upgrade:

https://github.blog/changelog/2024-09-25-actions-new-images-and-ubuntu-latest-changes/

To ensure we're caching consistently on the new image, switches from
using `ubuntu-latest` to `ubuntu-24.04`.

We can revert this change after October 30th, when the blog post says
the deployment is complete.